Do not wrap a function body with a block
So far I was thinking that the body of a function
should be a block of functions or data.
The problem with this approach is that a block
always returns a collection of items, so every function
would return an array of things.

This on the other hand means that I can't replace a thing
with a call to a function that returns this thing.

Somebodu could think that that's not a big problem,
because we can always call somethig like `first` on a result
of the function, and be able to refer to the actuall thing
that the function is supposed to return,
but how to implement `first`? It would be a function,
so it would return an array of things, insted of a thing itself.

This is a good change for another reason (than well... the whole thing
just won't work otherwise). Now we can write functions that return
their whole context (which I suspect will be super useful
when I get to loops).